Fixed syslink unhandled overrun interrupt
This commit fixes a bug, whereby an overrun of the syslink UART buffer was not handled in the interrupt routine, causing constant retriggering and reentry, stalling the system.

Overruns are caused when high CPU/interrupt load causes the syslink UART data buffer to not be processed fast enough.

This bug was fixed by:
 1. handling unhandled conditions in the ISR, and
 2. reordering/simplifying instructions in the ISR to reduce latency before data read.

As it is currently implemented, every received byte causes the CPU to be interrupted, and the byte to be sent to the syslink for processing (this is due to a variable packet size and hence unknown transaction length).

A nicer implementation would be to fix the packet length at its maximum (32 bytes), and use DMA for the entire RX transaction. This would entirely avert the problem of an overrun causing packet loss. I assert (without evidence) that the time required to interrupt the CPU every byte, is less than the overhead due to fixing the packet size at its maximum length.

Signed-off-by: Mike Hamer <mike@mikehamer.info>