Fixed crtpRxTask TODO
This commit fixes a TODO in the crtpRxTask.

In this task, packets are pushed to the communication port's length-1 queue. If this queue is full, the current functionality is to leave the old packet and discard the new one, however a TODO suggested replacing the old packet with the new one.

Since the queue is length 1, we use the FreeRTOS command xQueueOverwrite to do just this (replace the old packet with the new, or enqueue the new if the queue is empty).

Signed-off-by: Mike Hamer <mike@mikehamer.info>