add rotation direction heuristic, rem integrators
Added a heuristic for determining whether to keep rotating in the same direction and do a full rotation, or whether to try and halt the rotation and go back the other way. This is needed to be able to command a flip without using body-rate feed-forwards.

Removed the integrals because I wasn't using them, and when I did, they caused me issues.

Signed-off-by: Mike Hamer <mike@mikehamer.info>